requirements.txt
# Core Data Processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2

# Kafka
kafka-python==2.0.2

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Utilities
python-dateutil==2.8.2

# Optional (for notebooks)
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
